CA-118 split render
